requirements.txt
django-rest-knox==4.1.0
django==3.1.1
djangorestframework==3.11.1